% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiTVMM.R
\name{guiTVMM}
\alias{guiTVMM}
\title{A graphical user interface (GUI) of the package TVMM}
\usage{
guiTVMM(gui = TRUE)
}
\arguments{
\item{gui}{Logical argument, \code{TRUE} or \code{FALSE}. The default is \code{TRUE}}
}
\value{
\code{guiTVMM} A graphical user interface (GUI) for performing tests on the vector of multivariate population means.
}
\description{
A graphical user interface (GUI) of the package TVMM to perform more general hypothesis tests on the vector of multivariate population means.


\url{https://rpubs.com/Henriqueufla/617206} for more details.
}
\examples{
library(TVMM)
if(interactive()){
  guiTVMM(gui=FALSE)
}

}
\references{
Henrique J. P. Alves & Daniel F. Ferreira (2019): Proposition of new alternative tests adapted to the traditional T2 test, Communications in Statistics - Simulation and Computation, DOI: 10.1080/03610918.2019.1693596
}
